spdm processing
