+++ HIRS_Utils/src/main/java/hirs/utils/tpm/eventlog/events/DeviceSecurityEventData2.java
@@ -1,0 +1,13 @@
+package hirs.utils.tpm.eventlog.events;
+
+public class DeviceSecurityEventData2 extends DeviceSecurityEventDataBase {
+
+    /**
+     * DeviceSecurityEventData2 Constructor.
+     *
+     * @param dSEDbytes byte array holding the DeviceSecurityEventData.
+     */
+    public DeviceSecurityEventData2(final byte[] dSEDbytes) {
+
+    }
+}
