+++ HIRS_Utils/src/main/java/hirs/utils/tpm/eventlog/events/DeviceSecurityEventDataHeader2.java
@@ -1,0 +1,8 @@
+package hirs.utils.tpm.eventlog.events;
+
+public class DeviceSecurityEventDataHeader2 extends DeviceSecurityEventDataHeaderBase {
+
+    public DeviceSecurityEventDataHeader2(final byte[] dSEDbytes) {
+
+    }
+}
